refactor: comprehensive buffer architecture overhaul and optimizations
- Centralized all buffer-related utilities into `Shared.Buffers` namespace.
- Decoupled memory management via `ISlabAllocator` and `DefaultSlabAllocator`.
- Implemented `IBufferWriter<byte>` in `SnapshotBuffer` for standard .NET integration.
- Updated `BitReader` to natively support `ReadOnlySequence<byte>` and zero-copy multi-segment reading.
- Optimized `BitWriter` and `BitReader` with `BinaryPrimitives` fast paths and efficient LEB128/ZigZag encoding.
- Introduced `SlabList` helper to consolidate slab management and base offset tracking.
- Standardized diagnostics via `IBuffer` interface and `IDiagnosticBus` integration.
- Enhanced thread-safety in `DoubleBuffer` and `RingBuffer`.
- Expanded test suite with `BufferArchitectureTests.cs` to verify new architectural layers.
- Resolved multiple call-site and internal logic issues following the refactor.

+++ Engine/Shared/Networking/FieldHandlers/ComponentsFieldHandler.cs
@@ -69,10 +69,10 @@ public class ComponentsFieldHandler : INetworkFieldHandler
 
             if (finder.Found != null)
             {
-                int startBits = reader.BitsRead;
+                long startBits = reader.BitsRead;
                 finder.Found.ReadState(ref reader);
-                int actualRead = reader.BitsRead - startBits;
-                if (actualRead < payloadBits) reader.SkipBits(payloadBits - actualRead);
+                long actualRead = reader.BitsRead - startBits;
+                if (actualRead < payloadBits) reader.SkipBits((int)(payloadBits - actualRead));
             }
             else
             {
